requirements.txt
flask==2.3.2
torch==2.1.0
torchvision==0.16.0
timm==0.9.8
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.1.0
